agent/python/requirements.txt
livekit-agents>=0.7.0
livekit-plugins-openai>=0.7.0
livekit-plugins-silero>=0.7.0
livekit-plugins-deepgram>=0.7.0
aiohttp>=3.10.0
